feat: Implement fetching and retrieving today's weather forecast
This commit introduces the functionality to fetch weather forecasts for a city from a remote source, persist it locally, and retrieve today's weather forecast.

-   **domain/src/main/java/com/ab/domain/usecases/:**
    -   Added `FetchWeatherForecastByCityNameUseCase` to fetch weather forecasts and save them locally.
    -   Added `GetTodayWeatherForecastByCityUseCase` to retrieve today's weather forecast for a given city.
-   **domain/src/main/java/com/ab/domain/mapper/:**
    -   Added `CityMapper` to map between `CityDto` and `CityEntity`.
    -   Added `TemperatureMapper` to map between `TemperatureDto` and `TemperatureEntity`.
    -   Added `WeatherConditionMapper` to map between `WeatherConditionDto` and `WeatherConditionEntity`.
    -   Added `WeatherForecastMapper` to map between `WeatherForecastDto` and `WeatherForecastEntity`.
-   **domain/src/main/java/com/ab/domain/model/model/:**
    - Added model classes `WeatherCondition`, `Temperature`, `WeatherForecast`, `CityWeatherForecast`, `TodayWeatherForecast`.
-   **domain/src/main/java/com/ab/domain/repository/:**
    - Modified `WeatherForecastRepository` to include `getCityWeatherForecastByName` method.
-   **domain/src/main/java/com/ab/domain/model/entity/:**
    - Added `WeatherForecastDetailsEntityRel` and `CityWithWeatherForecastDetailsEntityRel` for relation.
    - Modified `CityEntity` changed population type to long.
-   **data/src/main/java/com/ab/data/repository/:**
    - Implemented `getCityWeatherForecastByName` method in `WeatherForecastRepositoryImpl`.
-   **data/src/main/java/com/ab/data/source/local/datasource/city/:**
    -   Modified `CityLocalDataSource` to include `getAllWithRelations` and `getCityWeatherForecastByName` methods.
    - Added the implementation for `getAllWithRelations` and `getCityWeatherForecastByName` in `CityLocalDataSourceImpl`.
-   **data/src/main/java/com/ab/data/source/local/db/dao/:**
    -   Added `getAllWithRelations` and `getCityWeatherForecastByName` to `CityDao`.
- **core/src/main/res/values/:**
    - Add new strings `no_city_found_with_this_name` and `no_weather_forecast_today` for error cases.
-   Modified `WeatherForecastDto` dateTimestamp property to be non-nullable.
- Moved relation object to `domain/src/main/java/com/ab/domain/model/entity/relations/`

+++ domain/src/main/java/com/ab/domain/model/entity/CityEntity.kt
@@ -12,6 +12,6 @@ data class CityEntity(
     val longitude: Double,
     val latitude: Double,
     val country: String,
-    val population: Int,
+    val population: Long,
     val timezone: Int
 )
